:construction: Separação entre index.js e firestore.js
